pyproject.toml
[project]
name = "next-showing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.0",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "uvicorn>=0.44.0",
]
